Drag drop works and code is rearranged
